requirements.txt
asyncio-mqtt==0.12.1
paho-mqtt==1.6.1
pymongo==4.2.0

















[loggers]
keys=root,tanka

[logger_root]
level=DEBUG
handlers=file_handler


[logger_tanka]
level=DEBUG
qualname=tanka
handlers=file_handler

[formatters]
keys=formatter

[handlers]
keys=file_handler

[handler_file_handler]
class=logging.handlers.TimedRotatingFileHandler
level=INFO
formatter=formatter
args=('sdk.log', 'W5', 0, 4)

[formatter_formatter]
format={"date":"%(asctime)s","level":"%(levelname)s","file":"%(filename)s","lineno":"%(lineno)d", "msg": "%(message)s"}
datefmt=%Y-%m-%d %H:%M:%S